+++ controller/controller.go
@@ -220,17 +220,17 @@ func (s *Controller) IncreaseTime() {
 
 // Move simply moves the stepper by the specified number of increments
 func (s *Controller) Move(n int32) {
-	rawMove := float32(n*int32(s.calibrationCfg.StepsPerIncrement)) + s.remainder
+	rawMove := float32(n)*s.calibrationCfg.StepsPerIncrement + s.remainder
 
 	move := int32(rawMove + 0.5)
 	s.remainder = rawMove - float32(move)
 
 	// add or subtract backlash steps based on direction change
 	if s.lastDirection < 0 && move > 0 {
-		move += int32(s.calibrationCfg.BacklashSteps)
+		move += s.calibrationCfg.BacklashSteps
 		s.lastDirection = +1
 	} else if s.lastDirection > 0 && move < 0 {
-		move -= int32(s.calibrationCfg.BacklashSteps)
+		move -= s.calibrationCfg.BacklashSteps
 		s.lastDirection = -1
 	}
